requirements.txt
django==4.2
aiogram
python-dotenv